+++ shared/include/shared/player.h
@@ -3,7 +3,7 @@
 
 struct player
 {
-    int map_index;
+    size_t map_index;
     float pos_x;
     float pos_y;
     float vel_x;
@@ -12,7 +12,7 @@ struct player
     float acc_y;
 };
 
-void player_init(struct player *player, int map_index);
+void player_init(struct player *player, size_t map_index);
 void player_accelerate(struct player *player, struct map *map, float delta_time);
 void player_attack(struct player *player, struct map *map);
